refactor: ディレクトリ構成改善
