+++ backend/src/auth/auth.service.ts
@@ -7,7 +7,7 @@ import { MailService } from '@src/mail/mail.service'
 import { PendingEmailChangeService } from '@src/pendingEmailChange/pendingEmailChange.service'
 import { PrismaService } from '@src/prisma/prisma.service'
 import { EmailAlreadyExistsException } from '@src/user/errors/emailAlreadyExists.exception'
-import { User } from '@src/user/models/user.model'
+import { User } from '@src/user/user.model'
 import { UserService } from '@src/user/user.service'
 import { EmailChangeRequestInput } from './dto/emailChangeRequest.input'
 import { SignInResponse } from './dto/signIn.response'
