+++ backend/src/auth/dto/signIn.response.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql'
 
-import { User } from '@src/user/models/user.model'
+import { User } from '@src/user/user.model'
 
 @ObjectType()
 export class SignInResponse {
