+++ backend/src/auth/strategies/local.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common'
 import { PassportStrategy } from '@nestjs/passport'
 import { Strategy } from 'passport-local'
 
-import { User } from '@src/user/models/user.model'
+import { User } from '@src/user/user.model'
 import { AuthService } from '../auth.service'
 import { CustomUnauthorizedException } from '../errors/customUnauthorized.exception'
