+++ backend/src/commentFavorite/commentFavorite.resolver.ts
@@ -3,9 +3,9 @@ import { Args, Int, Mutation, Resolver } from '@nestjs/graphql'
 
 import { AuthenticatedUser } from '@src/auth/decorators/currentUser.decorator'
 import { JwtAuthGuard } from '@src/auth/guards/jwtAuth.guard'
-import { User } from '@src/user/models/user.model'
+import { User } from '@src/user/user.model'
+import { CommentFavorite } from './commentFavorite.model'
 import { CommentFavoriteService } from './commentFavorite.service'
-import { CommentFavorite } from './models/commentFavorite.model'
 
 @Resolver()
 export class CommentFavoriteResolver {
