+++ backend/src/ideaFile/ideaFile.model.ts
@@ -1,7 +1,7 @@
 import { Field, Int, ObjectType } from '@nestjs/graphql'
 
-import { Idea } from '@src/idea/models/idea.model'
-import { User } from '@src/user/models/user.model'
+import { Idea } from '@src/idea/idea.model'
+import { User } from '@src/user/user.model'
 
 @ObjectType()
 export class IdeaFileRelations {
